Delete User Profile Utility

+++ frontend/src/app/components/product/product.component.ts
@@ -103,7 +103,7 @@ export class ProductComponent implements OnInit {
 
     dialogRef.afterClosed().subscribe(result => {
         console.log('Dialog Closed');
-        if (result.event === 'Delete') {
+        if (result !== undefined && result.event === 'Delete') {
             this.delete();
         }
     });
